@@ -11,9 +11,11 @@ group :test do
   gem 'minitest-hooks'
   gem 'simplecov', '0.13.0', require: false
   gem 'simplecov-console', require: false
-  gem 'codeclimate-test-reporter', '1.0.8', require: nil
   gem 'test_after_commit'
   gem 'minitest-retry'
+  gem 'webmock'
+  gem 'mocha'
+  gem 'rails-controller-testing'
 end
 
 group :doc do
@@ -28,18 +30,15 @@ group :development do
   gem 'htmlentities', require: false
 end
 
-# Pender Client
-gem 'webmock'
-gem 'mocha'
 gem 'pender_client', git: 'https://github.com/meedan/pender-client.git', ref: '89c9072'
 gem 'sqlite3', require: false
 gem 'lograge'
-gem 'rails', '4.2.11.1'
+gem 'rails', '5.2.4.2'
 gem 'pg', '~> 0.15'
 gem 'pg_application_name'
 gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.1.0'
+gem 'coffee-rails'
 gem 'jquery-rails'
 gem 'turbolinks'
 gem 'jbuilder', '~> 2.0'
@@ -66,19 +65,18 @@ gem 'graphql-formatter'
 gem 'nokogiri', '1.10.8'
 gem 'puma'
 gem 'rack-cors', '1.0.6', require: 'rack/cors'
-gem 'sidekiq'
+gem 'sidekiq', '~> 5.0.0'
 gem 'sinatra'
 gem 'pusher'
 gem 'redis-namespace'
 gem 'file_validators'
-gem 'clamav-client'
 gem 'filesize'
 gem 'rest-client'
 gem 'rails_admin', '~> 1.3.0'
 gem 'transifex-interface-ruby', git: 'https://github.com/meedan/transifex-interface-ruby.git'
 gem 'http_accept_language', git: 'https://github.com/iain/http_accept_language.git'
 gem 'i18n-generators'
-gem 'rails-i18n', git: 'https://github.com/svenfuchs/rails-i18n.git', branch: 'rails-4-x', ref: 'f8c83ec'
+gem 'rails-i18n', git: 'https://github.com/svenfuchs/rails-i18n.git', branch: 'rails-5-x', ref: 'fa7e382'
 gem 'devise-i18n'
 gem 'ruby-cldr'
 gem 'twitter'
@@ -87,12 +85,12 @@ gem 'request_store'
 gem 'geojsonlint'
 gem 'jsonpath', '~> 0.8.8'
 gem 'graphql-batch'
-gem 'google-protobuf', '3.2.0'
+gem 'google-protobuf'
 gem 'concurrent-ruby', require: 'concurrent'
 gem 'rails_admin_material'
 gem 'apollo-tracing', '1.5.0'
 gem 'rubyzip', '1.3.0'
-gem 'rack-protection', '1.5.5'
+gem 'rack-protection', '~> 2.0.1'
 gem 'loofah', '2.3.1'
 gem 'deep_cloneable', '~> 2.3.0'
 gem 'rails-html-sanitizer', '1.0.4'
@@ -105,7 +103,6 @@ gem 'ffi', '1.9.24'
 gem 'google_drive'
 gem 'activerecord-import'
 gem 'redis-rails'
-gem 'activejob', '4.2.11.1'
 gem 'rack', '>= 1.6.11'
 gem 'jwt'
 gem 'smooch-api', git: 'https://github.com/smooch/smooch-ruby.git', ref: 'a2c1646'
@@ -131,9 +128,10 @@ gem 'execjs'
 gem 'carrierwave-video-thumbnailer'
 gem 'carrierwave_backgrounder', git: 'https://github.com/meedan/carrierwave_backgrounder'
 gem 'activerecord-redshift-adapter',  git: 'https://github.com/arp/activerecord-redshift-adapter.git', branch: "rails4-compatibility"
-gem 'partitioned', git: 'https://github.com/dkhofer/partitioned.git', branch: 'rails-4-2'
+gem 'partitioned'
 gem 'mimemagic', '~> 0.3.3'
 gem 'honeycomb-beeline'
 gem 'bumbler', require: false
-gem 'swagger-docs', '0.1.9'
+gem 'swagger-docs'
 gem 'ruby-progressbar', '~> 1.10', '>= 1.10.1'
+gem 'omniauth-rails_csrf_protection', '~> 0.1'
