@@ -107,7 +107,7 @@ module AnnotationBase
         annotated.skip_clear_cache = self.skip_clear_cache
         annotated.updated_at = Time.now
         annotated.disable_es_callbacks = true
-        ActiveRecord::Base.connection_pool.with_connection do
+        ApplicationRecord.connection_pool.with_connection do
           annotated.save!(validate: false)
         end
       end
