@@ -101,7 +101,7 @@ module UserInvitation
     end
 
     def is_invited?(team = nil)
-      return false unless ActiveRecord::Base.connection.column_exists?(:users, :invitation_token)
+      return false unless ApplicationRecord.connection.column_exists?(:users, :invitation_token)
       team ||= Team.current
       return true if self.invited_to_sign_up?
       tu = self.team_users.where(status: 'invited', team_id: team.id).where.not(invitation_token: nil).last unless team.nil?
