@@ -1,4 +1,4 @@
-class Contact < ActiveRecord::Base
+class Contact < ApplicationRecord
   belongs_to :team
   phony_normalize :phone, default_country_code: 'US'
   validates :phone, phony_plausible: true
