@@ -1,4 +1,4 @@
-class Media < ActiveRecord::Base
+class Media < ApplicationRecord
   self.inheritance_column = :type
 
   attr_accessor :project_id, :project_object, :disable_es_callbacks
