@@ -1,4 +1,4 @@
-class TeamUser < ActiveRecord::Base
+class TeamUser < ApplicationRecord
 
   belongs_to :team
   belongs_to :user
