@@ -249,7 +249,7 @@ class Version < Partitioned::ByForeignKey
       if pa
         return unless pa.respond_to?(:cached_annotations_count)
         count = pa.cached_annotations_count + value
-        ActiveRecord::Base.connection_pool.with_connection do
+        ApplicationRecord.connection_pool.with_connection do
           pa.update_columns(cached_annotations_count: count)
         end
       end
