@@ -1,5 +1,5 @@
 class RenameBotsTable < ActiveRecord::Migration
   def change
-    rename_table(:bots, :bot_bots) if ActiveRecord::Base.connection.table_exists?(:bots)
+    rename_table(:bots, :bot_bots) if ApplicationRecord.connection.table_exists?(:bots)
   end
 end
