@@ -3,7 +3,7 @@ require 'error_codes'
 class ApplicationController < ActionController::Base
   include HttpAcceptLanguage::AutoLocale
 
-  protect_from_forgery with: :exception
+  protect_from_forgery with: :exception, prepend: true
   skip_before_filter :verify_authenticity_token
 
   before_action :configure_permitted_parameters, if: :devise_controller?
